@@ -1,15 +1,24 @@
 {
   "name": "pretty-quick",
-  "version": "0.0.0-development",
+  "version": "3.1.3",
   "description": "Get Pretty Quick",
   "repository": "azz/pretty-quick",
   "author": "Lucas Azzola <@azz>",
-  "main": "./dist",
-  "bin": "./bin/pretty-quick.js",
+  "maintainers": [
+    "JounQin (https://www.1stG.me) <admin@1stg.me>"
+  ],
   "license": "MIT",
+  "packageManager": "yarn@2.4.3",
   "engines": {
     "node": ">=10.13"
   },
+  "bin": "./bin/pretty-quick.js",
+  "main": "./dist",
+  "files": [
+    "bin",
+    "dist",
+    "img"
+  ],
   "keywords": [
     "git",
     "mercurial",
@@ -21,48 +30,53 @@
     "vcs",
     "precommit"
   ],
-  "files": [
-    "bin",
-    "dist",
-    "img"
-  ],
-  "dependencies": {
-    "chalk": "^3.0.0",
-    "execa": "^4.0.0",
-    "find-up": "^4.1.0",
-    "ignore": "^5.1.4",
-    "mri": "^1.1.5",
-    "multimatch": "^4.0.0"
-  },
   "scripts": {
-    "prepublishOnly": "yarn build",
-    "pretty-quick": "./bin/pretty-quick.js",
     "build": "babel src -d dist --copy-files --no-copy-ignored --ignore '**/__tests__/*.js'",
-    "test": "jest",
     "lint": "eslint . --ignore-path=.gitignore",
-    "semantic-release": "semantic-release"
+    "prepare": "simple-git-hooks",
+    "pretty-quick": "./bin/pretty-quick.js",
+    "release": "yarn build && clean-pkg-json && changeset publish",
+    "test": "jest"
   },
-  "husky": {
-    "hooks": {
-      "pre-commit": "./bin/pretty-quick.js --staged"
-    }
-  },
-  "prettier": "@azz/prettier-config",
   "peerDependencies": {
     "prettier": ">=2.0.0"
   },
+  "dependencies": {
+    "chalk": "^4.1.2",
+    "execa": "^4.1.0",
+    "find-up": "^4.1.0",
+    "ignore": "^5.3.0",
+    "mri": "^1.2.0",
+    "multimatch": "^5.0.0"
+  },
   "devDependencies": {
-    "@azz/prettier-config": "^1.0.0",
-    "@babel/cli": "^7.8.4",
-    "@babel/core": "^7.9.0",
-    "@babel/preset-env": "^7.9.0",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.10.1",
-    "eslint-plugin-jest": "^23.8.2",
-    "eslint-plugin-prettier": "^3.1.2",
-    "husky": "^4.2.3",
-    "jest": "^25.2.3",
-    "mock-fs": "^4.11.0",
-    "prettier": "2.0.2"
-  }
+    "@1stg/prettier-config": "3.9.2",
+    "@babel/cli": "^7.23.4",
+    "@babel/core": "^7.23.7",
+    "@babel/preset-env": "^7.23.8",
+    "@changesets/changelog-github": "^0.5.0",
+    "@changesets/cli": "^2.27.1",
+    "clean-pkg-json": "^1.2.0",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-jest": "^27.6.3",
+    "eslint-plugin-prettier": "^4.2.1",
+    "jest": "^26.6.3",
+    "mock-fs": "^4.14.0",
+    "npm-run-all": "^4.1.5",
+    "prettier": "2.8.8",
+    "simple-git-hooks": "^2.9.0",
+    "size-limit": "^11.0.1",
+    "size-limit-preset-node-lib": "^0.3.0"
+  },
+  "prettier": "@1stg/prettier-config",
+  "simple-git-hooks": {
+    "pre-commit": "./bin/pretty-quick.js --staged"
+  },
+  "size-limit": [
+    {
+      "path": "src/index.js",
+      "limit": "1KB"
+    }
+  ]
 }
