@@ -1,17 +1,17 @@
-import mock from 'mock-fs';
+import mock from 'mock-fs'
 
-import prettyQuick from '..';
+import prettyQuick from '..'
 
-jest.mock('execa');
+jest.mock('execa')
 
-afterEach(() => mock.restore());
+afterEach(() => mock.restore())
 
 test('throws an error when no vcs is found', () => {
   mock({
     'root/README.md': '',
-  });
+  })
 
   expect(() => prettyQuick('root')).toThrow(
     'Unable to detect a source control manager.',
-  );
-});
+  )
+})
