@@ -1,39 +1,39 @@
-import findUp from 'find-up';
-import execa from 'execa';
-import { dirname, join } from 'path';
-import * as fs from 'fs';
+import findUp from 'find-up'
+import execa from 'execa'
+import { dirname, join } from 'path'
+import * as fs from 'fs'
 
-export const name = 'git';
+export const name = 'git'
 
-export const detect = (directory) => {
+export const detect = directory => {
   if (fs.existsSync(join(directory, '.git'))) {
-    return directory;
+    return directory
   }
 
   const gitDirectory = findUp.sync('.git', {
     cwd: directory,
     type: 'directory',
-  });
+  })
   if (gitDirectory) {
-    return dirname(gitDirectory);
+    return dirname(gitDirectory)
   }
 
   const gitWorktreeFile = findUp.sync('.git', {
     cwd: directory,
     type: 'file',
-  });
+  })
 
   if (gitWorktreeFile) {
-    return dirname(gitWorktreeFile);
+    return dirname(gitWorktreeFile)
   }
-};
+}
 
 const runGit = (directory, args) =>
   execa.sync('git', args, {
     cwd: directory,
-  });
+  })
 
-const getLines = (execaResult) => execaResult.stdout.split('\n');
+const getLines = execaResult => execaResult.stdout.split('\n')
 
 export const getSinceRevision = (directory, { staged, branch }) => {
   try {
@@ -43,18 +43,18 @@ export const getSinceRevision = (directory, { staged, branch }) => {
           'merge-base',
           'HEAD',
           branch || 'master',
-        ]).stdout.trim();
-    return runGit(directory, ['rev-parse', '--short', revision]).stdout.trim();
+        ]).stdout.trim()
+    return runGit(directory, ['rev-parse', '--short', revision]).stdout.trim()
   } catch (error) {
     if (
       /HEAD/.test(error.message) ||
       (staged && /Needed a single revision/.test(error.message))
     ) {
-      return null;
+      return null
     }
-    throw error;
+    throw error
   }
-};
+}
 
 export const getChangedFiles = (directory, revision, staged) => {
   return [
@@ -75,13 +75,13 @@ export const getChangedFiles = (directory, revision, staged) => {
       : getLines(
           runGit(directory, ['ls-files', '--others', '--exclude-standard']),
         )),
-  ].filter(Boolean);
-};
+  ].filter(Boolean)
+}
 
-export const getUnstagedChangedFiles = (directory) => {
-  return getChangedFiles(directory, null, false);
-};
+export const getUnstagedChangedFiles = directory => {
+  return getChangedFiles(directory, null, false)
+}
 
 export const stageFile = (directory, file) => {
-  runGit(directory, ['add', file]);
-};
+  runGit(directory, ['add', file])
+}
