@@ -1,13 +1,13 @@
-import * as gitScm from './git';
-import * as hgScm from './hg';
+import * as gitScm from './git'
+import * as hgScm from './hg'
 
-const scms = [gitScm, hgScm];
+const scms = [gitScm, hgScm]
 
-export default (directory) => {
+export default directory => {
   for (const scm of scms) {
-    const rootDirectory = scm.detect(directory);
+    const rootDirectory = scm.detect(directory)
     if (rootDirectory) {
-      return Object.assign({ rootDirectory }, scm);
+      return Object.assign({ rootDirectory }, scm)
     }
   }
-};
+}
